finish update_deck route

+++ app/api/deck_routes.py
@@ -103,31 +103,33 @@ def update_deck():
 
 
 @deck_routes.route("/<int:id>", methods=["PATCH"])
+@login_required
 def update_deck(id):
     deck = Deck.query.get(id)
     cards_in_table = deck.to_dict_with_cards()
     form = CardForm()
-    # form["csrf_token"].data = request.cookies["csrf_token"]
+    form["csrf_token"].data = request.cookies["csrf_token"]
     data = request.json
-    print("FFFFF ===>" + str(data["cards"]))
-    print("FFFFF ===>" + str(cards_in_table))
 
-    for card in data["cards"]:
-        if int(card["id"]) == 0:
-            card = Card(question=card["question"], answer=card["answer"])
-            deck.cards.append(card)
-            db.session.add(card)
+    if deck.userId == current_user.id:
+        for card in data["cards"]:
+            if int(card["id"]) == 0:
+                card = Card(question=card["question"], answer=card["answer"])
+                deck.cards.append(card)
+                db.session.add(card)
+                db.session.commit()
+            elif cards_in_table[int(card["id"])]:
+                card_from_table = Card.query.get(int(card["id"]))
+                card_from_table.question = card["question"]
+                card_from_table.answer = card["answer"]
+                del cards_in_table[int(card["id"])]
+                db.session.commit()
+
+        for key in cards_in_table.keys():
+            card = Card.query.get(key)
+            db.session.delete(card)
             db.session.commit()
-        elif cards_in_table[int(card["id"])]:
-            card_from_table = Card.query.get(int(card["id"]))
-            card_from_table.question = card["question"]
-            card_from_table.answer = card["answer"]
-            del cards_in_table[int(card["id"])]
-            db.session.commit()
 
-    for key in cards_in_table.keys():
-        card = Card.query.get(key)
-        db.session.delete(card)
-        db.session.commit()
+        return deck.to_dict()
 
-    return deck.to_dict()
+    return {"errors": ["Unauthorized"]}
